Refactor SalesQuotationSchItemDtlDto, CreateSalesQuotationSchItemDtlRequest, and SalesQuotationSchItemDtl to use BigDecimal for price and cost fields for improved precision in financial calculations.

+++ src/main/java/com/asg/shipchandling/salesquotationsch/dto/request/CreateSalesQuotationSchItemDtlRequest.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,7 +17,7 @@ public class CreateSalesQuotationSchItemDtlRequest {
     
     private Long quantity;
     
-    private Long price;
+    private BigDecimal price;
     
     private Long discount;
     
@@ -28,7 +30,7 @@ public class CreateSalesQuotationSchItemDtlRequest {
     
     private Long adjQuantity;
     
-    private Long cost;
+    private BigDecimal cost;
     
     private Long lastRate1;
